Generic test runner: Migrate Checkbox tests (6/n)
[ghstack-poisoned]
