+++ packages/component-driver-mui-v5-test/__tests__/ButtonDriver.e2e.test.ts
@@ -1,7 +1,6 @@
 import { goto, playwrightGetTestEngine, playWrightTestAdapter } from '@atomic-testing/playwright';
 import { testRunner } from '@atomic-testing/test-runner';
-import { complexButtonTestSuite } from '../src/examples/button/Complex.example';
-import { iconAndLabelButtonTestSuite } from '../src/examples/button/IconAndLabel.example';
+import { complexButtonTestSuite, iconAndLabelButtonTestSuite } from '../src/examples';
 
 testRunner(iconAndLabelButtonTestSuite, playWrightTestAdapter, {
   // @ts-ignore
