+++ packages/component-driver-mui-v5-test/__tests__/Rating.dom.test.ts
@@ -1,7 +1,7 @@
 import { jestTestAdapter } from '@atomic-testing/jest';
 import { createTestEngine } from '@atomic-testing/react';
 import { testRunner } from '@atomic-testing/test-runner';
-import { basicRatingExample, ratingTestSuite } from '../src/examples/rating';
+import { basicRatingExample, ratingTestSuite } from '../src/examples';
 
 testRunner(ratingTestSuite, jestTestAdapter, {
   getTestEngine: (scenePart: typeof basicRatingExample.scene) => {
