+++ packages/component-driver-mui-v5-test/__tests__/Rating.e2e.test.ts
@@ -1,6 +1,6 @@
 import { goto, playwrightGetTestEngine, playWrightTestAdapter } from '@atomic-testing/playwright';
 import { testRunner } from '@atomic-testing/test-runner';
-import { ratingTestSuite } from '../src/examples/rating';
+import { ratingTestSuite } from '../src/examples';
 
 testRunner(ratingTestSuite, playWrightTestAdapter, {
   // @ts-ignore
