+++ packages/component-driver-mui-v5-test/src/examples/button/index.ts
@@ -1,6 +1,8 @@
 import { IExampleUnit, ScenePart } from '@atomic-testing/core';
-import { complexExample as complex } from './Complex.example';
-import { iconAndLabelExample as iconAndLabel } from './IconAndLabel.example';
+import { complexExample as complex } from './ComplexButton.example';
+import { iconAndLabelExample as iconAndLabel } from './IconAndLabelButton.example';
+export { complexButtonTestSuite } from './ComplexButton.example';
+export { iconAndLabelButtonTestSuite } from './IconAndLabelButton.example';
 
 export const iconAndLabelExample = iconAndLabel;
 export const complexExample = complex;
